Split Express app setup from server setup
Separates their concerns: middleware, routing, handlers, etc. belong to
the app while CLI argument parsing, ports, listening, etc. belong to the
server.
